requirements.txt
aiogram
Pillow
aiohttp